refactor: 요청, 출동 게시물 작성 로직 수정
