+++ src/main/java/com/example/wonderwoman/delivery/controller/DeliveryController.java
@@ -1,11 +1,10 @@
 package com.example.wonderwoman.delivery.controller;
 
 import com.example.wonderwoman.common.dto.NormalResponseDto;
-import com.example.wonderwoman.delivery.DeliveryPostResponseDto;
-import com.example.wonderwoman.delivery.DeliveryResponseDto;
 import com.example.wonderwoman.delivery.entity.DeliveryPost;
 import com.example.wonderwoman.delivery.request.DeliveryRequestDto;
 import com.example.wonderwoman.delivery.service.DeliveryService;
+import com.example.wonderwoman.login.CurrentUser;
 import com.example.wonderwoman.member.entity.Member;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
@@ -13,7 +12,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -27,7 +25,7 @@ public class DeliveryController {
 
     //딜리버리 게시글 작성
     @PostMapping("/post")
-    public ResponseEntity<NormalResponseDto> postDelivery(Member member, @RequestBody @Valid DeliveryRequestDto requestDto) {
+    public ResponseEntity<NormalResponseDto> postDelivery(@CurrentUser Member member, @RequestBody @Valid DeliveryRequestDto requestDto) {
         deliveryService.postDelivery(member, requestDto);
         return ResponseEntity.ok(NormalResponseDto.success());
     }
