--- src/main/java/com/example/wonderwoman/delivery/entity/DeliveryPostId.java
@@ -1,14 +1,0 @@
-package com.example.wonderwoman.delivery.entity;
-
-import jakarta.persistence.*;
-import lombok.Builder;
-
-import java.io.Serializable;
-
-@Embeddable
-public class DeliveryPostId implements Serializable {
-    private Long id;
-
-    private Long postId;
-
-}
