+++ src/main/java/com/example/wonderwoman/delivery/request/DeliveryRequestDto.java
@@ -23,17 +23,13 @@ public class DeliveryRequestDto {
 
     @NotBlank(message = "제목은 필수 입력값입니다.")
     private String postTitle;
-
-    @NotBlank(message = "요청 또는 출동을 선택하세요.")
+    
     private ReqType reqType;
 
-    @NotBlank(message = "개수는 필수 입력값입니다.")
     private int postNumber;
 
-    @NotBlank(message = "크기는 필수 입력값입니다.")
     private SanitarySize sanitarySize;
 
-    @NotBlank(message = "종류는 필수 입력값입니다.")
     private SanitaryType sanitaryType;
 
     private String postComment;
